src/app/modules/login/login-form/login-form.component.spec.ts
import { AuthService } from '@pl-core/_services/auth/auth.service';
/**
 *  ISSUES:
 *  - Cannot call fixture.detectChanges without a Material error at the
 *    spec level (inside an it() block), but works fine at the suite
 *    level (inside the beforeEach() block). This means we can't
 *    test the component's initial state before the OnInit hook
 */

import {ComponentFixture, TestBed, fakeAsync, tick, inject} from '@angular/core/testing';
import {By} from '@angular/platform-browser';
import {DebugElement, NO_ERRORS_SCHEMA} from '@angular/core';
import {Observable} from 'rxjs/Observable';
import {click} from '@pl/testing/utils';
import {RouterStub} from '@pl/testing/stubs';

// Imports for TestBed module metadata config
import {CommonModule} from '@angular/common';
import {CoreModule} from '@pl-core/core.module';
import {ReactiveFormsModule} from '@angular/forms';
import {MaterialModule} from '@angular/material';

import {TranslatePipe} from '@pl-core/_pipes';

// Import dependencies in name only, for the component under test
// Must configure in TestBed later
import {Router} from '@angular/router';
import {  FormBuilder, Validators, FormGroup } from '@angular/forms';
import {
  PaletteConfiguratorService,
  PaletteFetcherService,
  TranslateService,
  LoginService,
  LoadingService
} from '@pl-core/_services';
import {IAuthResponse} from '@pl-core/_interfaces';
import 'hammerjs';
import {AlertService} from '@pl-core/_services/alert/alert.service';
import {Subject} from 'rxjs/Subject';
import {LoginFormComponent} from './login-form.component';
import {RouterTestingModule} from "@angular/router/testing";
import {BrowserAnimationsModule} from '@angular/platform-browser/animations';

// --- Declarations and helpers ---
let comp: LoginFormComponent;
let fixture: ComponentFixture<LoginFormComponent>;
let de: DebugElement;
let page: Page;


class LoadingServiceSpy {
  public show = jasmine.createSpy('show').and.callFake(() => Observable.of(true));
  public hide = jasmine.createSpy('hide');
}

class TranslateServiceStub {

  constructor(){
    
  }
  public currentLang: string;

  public use(lang: string) {
    this.currentLang = lang;
  }

  public instant(key: string) {
    return 'translated';
  }

  public instantWithLang(key: string, lang: string) {
    return `translated to ${lang}`;
  }
}

class AlertServiceStub {
  private subject = new Subject<any>();
  private keepAfterNavigationChange = false;

  public success(message: string, keepAfterNavigationChange = false) {

    return ({type: 'success', text: message});
  }

  public error(message: string, keepAfterNavigationChange = false) {
    this.keepAfterNavigationChange = keepAfterNavigationChange;
    return {type: 'error', text: message};
  }

  public getMessage(): Observable<any> {
    return this.subject.asObservable();
  }
}

class LoginServiceStub {
  public getResources = jasmine.createSpy('getResources').and.callFake(
    () => Observable.of({
      logoURL: '#logo-url',
      logoALT: 'logo',
      slides: []
    })
  );
}

class AuthenticatedServiceSpy {
  public redirectUrl: string;

  public login(username: string, password: string): Observable<IAuthResponse> {
    return Observable.of({authenticated: true, error: 'None'});
  }

}

class UnAuthenticatedServiceSpy {
  public redirectUrl: string;

  public login(username: string, password: string): Observable<IAuthResponse> {
    return Observable.of({authenticated: false, error: 'Test Simulate Login Fail'});
  }
}

// Examples of spy stubs, where the methods of the actual class are
// stubbed by Jasmine spy objects rather than actual functions
class PaletteConfigSpy {
  public populateTenantColors = jasmine.createSpy('populateTenantColors');
}

class PaletteFetcherSpy {
  public getTenantPalette = jasmine.createSpy('getTenantPalette').and.callFake(
    () => Observable.of('palette')
  );
}

// Represents the component under test and aggregates all
// DOM elements/DebugElements/spies into one class
class Page {
  public onSubmitSpy: jasmine.Spy;

  // public carousel: DebugElement;
  public loginFormComponent: DebugElement;
  public loginForm: DebugElement;
  public logo: HTMLImageElement;
  public usernameInput: any = "username";
  public passwordInput: any = "password";
  public languageInput: any = "en";
  // public passwordInput: HTMLInputElement;
  // public languageInput: HTMLInputElement;
  // public usernameInput: HTMLInputElement;
  // forgotPassAnchor: HTMLAnchorElement;
  // changePassAnchor: HTMLAnchorElement;
  public loginBtn: HTMLButtonElement;

  constructor() {
    this.onSubmitSpy = spyOn(comp, 'onSubmit').and.callThrough();

    // // this.carousel = de.query(By.css('pl-carousel'));
    // // this.loginFormComponent = de.query(By.css('login-form'));
    // this.loginForm = de.query(By.css('form'));
    // this.logo = de.query(By.css('#logo')).nativeElement;
    // this.usernameInput = de.query(By.css('[formControlName=username]')).nativeElement;
    // this.passwordInput = de.query(By.css('[formControlName=password]')).nativeElement;
    // this.languageInput = de.query(By.css('md-select')).nativeElement;
    // // this.forgotPassAnchor = de.query(By.css('#forgot-pass')).nativeElement;
    // // this.changePassAnchor = de.query(By.css('#change-pass')).nativeElement;
    // this.loginBtn = de.query(By.css('.login-button')).nativeElement;
  }
}

function createComponent(): void {
  fixture = TestBed.createComponent(LoginFormComponent);
  comp = fixture.componentInstance;
  de = fixture.debugElement;
  page = new Page();

  // First change detection round triggers ngOnInit() and all
  // data/property/event bindings in the template
  fixture.detectChanges();
}
// --- End Declarations and helpers ---

// --- Tests ---
describe('LoginFormComponent', () => {
  beforeEach(() => {
    // TranslatePipe not stubbed but it uses the stubbed TranslateService
    TestBed.configureTestingModule({
      imports: [
        CoreModule,
        RouterTestingModule,
        BrowserAnimationsModule
      ],
      declarations: [LoginFormComponent],
      schemas: [NO_ERRORS_SCHEMA], // ignore <pl-carousel>
      providers: [
        {provide: AuthService, useClass: AuthenticatedServiceSpy},
        // {provide: Router, useClass: RouterStub},
        FormBuilder,
        {provide: LoadingService, useClass: LoadingServiceSpy},
        {provide: TranslateService, useClass: TranslateServiceStub},
        {provide: LoginService, useClass: LoginServiceStub},
        {provide: AlertService, useClass: AlertServiceStub}
      ]
    });

    createComponent();
  });

  describe('Unit tests - no embedded components', () => {

    it('buildForm() should instantiate a fresh FormGroup', () => {

      expect(comp.loginForm.contains('username')).toBe(true);
      console.log(comp.loginForm, "comp.loginForm.contains('username')")
      expect(comp.loginForm.contains('password')).toBe(true);
      expect(comp.loginForm.contains('language')).toBe(true);
      expect(comp.loginForm.valid).toBe(false);
      expect(comp.loginForm.pristine).toBe(true);
      expect(comp.loginForm.untouched).toBe(true);
    });

    it('onSubmit() should login if the form is valid', 
    inject([AuthService, LoadingService], 
      (authservice: AuthService, loadingservice: LoadingService) => {
      // page.usernameInput = 'username';
      // page.passwordInput = 'password';
      // let loginSpy = spyOn(comp, 'login');

      // Simulate user input from the DOM
      // page.usernameInput.value = 'username';
      // page.passwordInput.value = 'password';
      // page.usernameInput.dispatchEvent(new Event('input'));
      // page.passwordInput.dispatchEvent(new Event('input'));

      // // Let Angular know DOM has changed
      // fixture.detectChanges();
      let response = {
        authenticated : true,
        error: 'error message'
      }
      comp.loginForm.controls.username.setValue('jayanth');  
      comp.loginForm.controls.password.setValue('jayanth');      
      comp.loginForm.controls.language.setValue('en');
      this.redirectUrl = '';
      let spyon = spyOn(authservice, 'login')
      spyon.and.returnValue(Observable.of(response));
      comp.onSubmit();
      expect(loadingservice.hide).toHaveBeenCalled();
    }));

    it('Validation for Invalid user',
    inject([AuthService, LoadingService, AlertService],
      (authservice: AuthService, loadingservice: LoadingService, alertservice: AlertService) => {
      let response = {
        authenticated : false,
        error: 'error message'
      }
      comp.loginForm.controls.username.setValue('username');
      comp.loginForm.controls.password.setValue('password');
      comp.loginForm.controls.language.setValue('cn');
      fixture.detectChanges();
      let spyon = spyOn(authservice, 'login');
      spyon.and.returnValue(Observable.of(response));
      spyOn(alertservice, 'error');
      comp.onSubmit();
      expect(comp.loginError).toBe(true);
      expect(alertservice.error).toHaveBeenCalled();
      expect(loadingservice.hide).toHaveBeenCalled();
    }));

    it('Show error If there is no response',
    inject([AuthService, LoadingService, AlertService],
      (authservice: AuthService, loadingservice: LoadingService, alertservice: AlertService) => {
      let error = 'error';
      comp.loginForm.controls.username.setValue('username');
      comp.loginForm.controls.password.setValue('password');
      comp.loginForm.controls.language.setValue('cn');
      fixture.detectChanges();
      let spyon = spyOn(authservice, 'login');
      spyon.and.returnValue(Observable.of(error));
      spyOn(alertservice,'error');
      comp.onSubmit();
      expect(alertservice.error).toHaveBeenCalled();
      expect(loadingservice.hide).toHaveBeenCalled();
    }));

    it('check disableErrorMsg() function', () => {
      let loginError: any = true;
      comp.disableErrorMsg();
      expect(comp.loginError).toBe(false);
    })

    // it('logout() should delegate to AuthService', () => {
    //   // logout() unsubscribes from tenantPalette, so we need to
    //   // initialise the subscription first in initalisePalette()
    //   comp.initialisePalette();
    //   let authService = de.injector.get(AuthService);
    //   let logoutSpy = spyOn(authService, 'logout');

    //   comp.logout();
    //   expect(logoutSpy.calls.any()).toBe(true);
    // });

    describe('Login panel', () => {
      xit('should have a logo', () => {
        expect(page.logo).toBeDefined();
        console.log(page.logo, page, "Login Panel")
        expect(page.logo.alt).toEqual(comp.logoALT);
        expect(page.logo.src).toContain(comp.logoURL);
      });

      it('should have username, password, and language fields', () => {
        expect(page.usernameInput).toBeDefined();
        expect(page.passwordInput).toBeDefined();
        expect(page.languageInput).toBeDefined();
      });

      it('should have at least one language (English)', () => {
        expect(comp.languages.length).toBeGreaterThan(0);
        expect(comp.languages).toContain({
          label: 'English', value: 'en'
        });
      });

      xit('should change the language when one is selected', () => {
        // Mock user input with in-component change to the
        // language FormControl's value
        comp.loginForm.get('language').setValue('zh');
        fixture.detectChanges();
        let translateService = de.injector.get(TranslateService);
        let stubInstance = new TranslateServiceStub();
        console.log('stubInstance : ', stubInstance.currentLang)
        expect(stubInstance.currentLang).toEqual('zh');
// Deeper tests into <md-select> left to Angular Material
      });

      // it('should change the language when one is selected', () => {
      //   // Mock user input with in-component change to the
      //   // language FormControl's value
      //   comp.loginForm.get('language').setValue('zh');

      //   let translateService = de.injector.get(TranslateService);
      //   let stubInstance = new TranslateServiceStub()
      //     console.log('stubInstance : ', stubInstance.use())
      //   expect(translateService.use()).toEqual('zh');
      //   // Deeper tests into <md-select> left to Angular Material
      // });

      //   it('should have links to change and reset password', () => {
      //     expect(page.forgotPassAnchor).toBeDefined();
      //     expect(page.changePassAnchor).toBeDefined();
      //   });

      xit('should submit the form when a submit event is raised', () => {
        page.loginForm.triggerEventHandler('ngSubmit', null);
        expect(page.onSubmitSpy.calls.any()).toBe(true);
        expect(comp.loginForm.valid).toBe(false);
      });

      xit('should submit the form when login button is clicked', () => {
        click(page.loginBtn);
        expect(page.onSubmitSpy.calls.any()).toBe(true);
        expect(comp.loginForm.valid).toBe(false);
      });
    });

    // describe('login()', () => {
    //   let authService: AuthService;
    //   let router: Router;
    //   let routerSpy: jasmine.Spy;

    //   let mockUrl = 'mock/url';

    //   beforeEach(() => {
    //     authService = de.injector.get(AuthService);
    //     router = de.injector.get(Router);
    //     routerSpy = spyOn(router, 'navigate');
    //   });

    //   it('should redirect to a stored URL if available', () => {
    //     authService.isAuthenticated = true;
    //     authService.redirectUrl = mockUrl;

    //     comp.login();
    //     expect(routerSpy).toHaveBeenCalledWith([mockUrl]);
    //   });

    //   it('should redirect to the dashboard if not', () => {
    //     authService.isAuthenticated = true;

    //     comp.login();
    //     expect(routerSpy).toHaveBeenCalledWith(['/dashboard']);
    //   });

    //   it('should initialise the palette', () => {
    //     let initPaletteSpy = spyOn(comp, 'initialisePalette');
    //     authService.isAuthenticated = true;

    //     comp.login();
    //     expect(initPaletteSpy.calls.any()).toBe(true);
    //   });

    //   it('should not do anything if not authenticated', () => {
    //     comp.login();
    //     expect(authService.isAuthenticated).toBe(false);
    //     expect(routerSpy.calls.any()).toBe(false);
    //   });
    // });

  });

});
